requirements.txt
tqdm==4.37.0
biopython==1.74